% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idwt.R
\name{iwt}
\alias{iwt}
\title{Inverse discrete wavelet transform}
\usage{
iwt(x, wf = "la8", J = NULL)
}
\arguments{
\item{x}{a 1, 2, or 3 dimensional data array. The size of each dimension must
be dyadic.}

\item{wf}{the type of wavelet family used. See R-package waveslim for options.}

\item{J}{is the level (depth) of the decomposition. For default \code{NULL} the max
depth is used  making \code{iwt(x)} equal to multiplying \code{x} with the
inverse of corresponding wavelet matrix.}
}
\value{
\item{...}{An array with dimensions equal to those of \code{x}.}
}
\description{
This function performs a level J decomposition of the input
array (1d, 2d, or 3d) using the pyramid algorithm (Mallat 1989).
}
\details{
This is a C++/R wrapper function for a C implementation of the
inverse discrete wavelet transform by Brandon Whitcher, Rigorous Analytics Ltd,
 licensed under the BSD 3 license
https://cran.r-project.org/web/licenses/BSD_3_clause, see the Waveslim package;
Percival and Walden (2000); Gencay, Selcuk and Whitcher (2001).

Given a data array (1d, 2d or 3d) with dyadic
sizes this transform is computed efficiently via the pyramid
algorithm see Mallat (1989).

This functionality is used in the computations underlying \code{\link{softmaximin}}
to perform multiplications involving the wavelet (design) matrix efficiently.
}
\examples{
###1d
x <- as.matrix(rnorm(2^3))
range(x - iwt(wt(x)))

###2d
x <- matrix(rnorm(2^(3 + 4)), 2^3, 2^4)
range(x - iwt(wt(x)))

###3d
x <- array(rnorm(2^(3 + 4 + 5)), c(2^3, 2^4, 2^5))
range(x - iwt(wt(x)))

}
\references{
Gencay, R., F. Selcuk and B. Whitcher (2001) An Introduction to Wavelets and
Other Filtering Methods in Finance and Economics, Academic Press.

Mallat, S. G. (1989) A theory for multiresolution signal decomposition: the
wavelet representation, IEEE Transactions on Pattern Analysis and Machine
Intelligence, 11, No. 7, 674-693.

Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series
Analysis, Cambridge University Press.
}
\author{
Adam Lund, Brandon Whitcher
}
